Add server-rendered lorebook page
